requirements.txt
python-dotenv==1.0.0
openai==1.4.0
deepgram-sdk==2.12.0
pyaudio==0.2.14
webrtcvad-wheels==2.0.12
elevenlabs==0.2.27
pygame==2.5.2
openmeteo-requests==1.1.0
requests-cache==0.9.8
retry-requests==1.0.0
pandas==2.2.0
taipy==4.1.0
geopy==2.4.0
rhasspy-silence==0.4.0